calendar-service/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aio-pika>=9.0.0
aioredis>=2.0.0
python-json-logger>=2.0.0
alembic>=1.9.0 